backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
openai
google-generativeai
python-dotenv
pillow
geopy
websockets
